Migrate organization_bootstrap tests
